Refactor filters, locations, views, tests
