+++ src/components/cards/filters/EffectFilter.js
