+++ src/components/cards/filters/StrengthFilter.js
